devanagari-api/app/requirements.txt
Flask==3.1.0
tensorflow==2.19.0
numpy==2.1.3
Pillow==11.1.0
Werkzeug==3.1.3
opencv-python-headless==4.11.0.86
gunicorn